Fix altitude handling
Altitude was hardcoded to be 420 (lol). This goes ahead and does a
simple check to see if there is a "m" in the altitude string, and if
so, coverts the numerical part to feet. If "ft" or no units are indicated
we assume the value is given in feet.

We always send feet. That seems fine. But now you can configure either.